+++ guang-cli/packages/utils/src/index.ts
@@ -1,0 +1,7 @@
+import NpmPackage from "./NpmPackage.js";
+import * as versionUtils from './versionUtils.js';
+
+export {
+    NpmPackage,
+    versionUtils
+}