+++ guang-cli/packages/cli/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @guang-cli/cli
 
+## 0.0.7
+
+### Patch Changes
+
+- @guang-cli/create@0.3.1
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @guang-cli/create@0.3.0
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @guang-cli/create@0.2.0
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @guang-cli/create@0.1.0
+
+## 0.0.3
+
+### Patch Changes
+
+- @guang-cli/create@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes