+++ guang-cli/packages/template-react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @guang-cli/template-react
 
+## 1.3.0
+
+### Minor Changes
+
+- eslint
+
+## 1.2.0
+
+### Minor Changes
+
+- ejs template
+
 ## 1.1.0
 
 ### Minor Changes