+++ guang-cli/packages/template-vue/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @guang-cli/template-vue
 
+## 1.2.0
+
+### Minor Changes
+
+- ejs template
+
 ## 1.1.0
 
 ### Minor Changes